requirements.txt
asgiref==3.2.10
Django==3.1.3
django-crispy-forms==1.9.2
django-google-maps==0.12.2
django-summernote==0.8.11.6
django-useraudit==1.7.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
